refactor(server): split command payload rules
